python: update to reflect latest API change

+++ bindings/python/capstone/capstone.py
@@ -118,7 +118,7 @@ class _cs_arch(ctypes.Union):
 class _cs_insn(ctypes.Structure):
     _fields_ = (
         ('id', ctypes.c_uint),
-        ('address', ctypes.c_size_t),
+        ('address', ctypes.c_uint64),
         ('size', ctypes.c_uint16),
         ('mnemonic', ctypes.c_char * 32),
         ('op_str', ctypes.c_char * 96),
@@ -138,7 +138,7 @@ def _setup_prototype(lib, fname, restype, *argtypes):
 
 _setup_prototype(_cs, "cs_open", ctypes.c_int, ctypes.c_uint, ctypes.c_uint, ctypes.POINTER(ctypes.c_size_t))
 _setup_prototype(_cs, "cs_disasm_dyn", ctypes.c_size_t, ctypes.c_size_t, ctypes.POINTER(ctypes.c_char), ctypes.c_size_t, \
-        ctypes.c_size_t, ctypes.c_size_t, ctypes.POINTER(ctypes.POINTER(_cs_insn)))
+        ctypes.c_uint64, ctypes.c_size_t, ctypes.POINTER(ctypes.POINTER(_cs_insn)))
 _setup_prototype(_cs, "cs_free", None, ctypes.c_void_p)
 _setup_prototype(_cs, "cs_close", ctypes.c_int, ctypes.POINTER(ctypes.c_size_t))
 _setup_prototype(_cs, "cs_reg_name", ctypes.c_char_p, ctypes.c_size_t, ctypes.c_uint)
